Funcionando íntegramente las líneas de operación Person y Experience con un solo controller.

+++ src/main/java/com/portfolio/MyPortfolio8/service/IExperienceService.java
@@ -7,8 +7,8 @@ public interface IExperienceService {
     
     public void createExperience(Experience exp);
     public List<Experience> seeListExperience();
-    public void deleteExperience(int id);
-    public Experience findExperience(int id);
+    public void deleteExperience(Long id);
+    public Experience findExperience(Long id);
     public Experience editExperience(Experience exp);
     
 }
